refactor test suite.
