+++ docs/src/stopping_criterion.md
