+++ test/Project.toml
@@ -1,0 +1,18 @@
+[deps]
+AlgorithmsInterface = "d1e3940c-cd12-4505-8585-b0a4b322527d"
+AlgorithmsInterfaceTestSuite = "876e16e5-0a55-4659-a4bc-e03625482f1c"
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[sources]
+AlgorithmsInterface = {path = "../"}
+AlgorithmsInterfaceTestSuite = {path = "AlgorithmsInterfaceTestSuite.jl"}
+
+[compat]
+Aqua = "0.8"
+SafeTestsets = "0.1"
+Test = "1.10"
+julia = "1.10"
